app/requirements.txt
Flask==0.10.1
mysql-connector-python==8.0.13